- database update
- removed vessel type check from get application documents

+++ Bunkering.Core/ViewModels/AppDepotViewModel.cs
@@ -7,6 +7,6 @@ namespace Bunkering.Core.ViewModels
         public int? AppId { get; set; }
         public string Name { get; set; }
         public int ProductId { get; set; }
-        public int Volume {  get; set; }
+        public decimal Volume {  get; set; }
     }
 }
